refactor / cleanup authz sample
